requirements.txt
Django>=5.1,<5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
bcrypt==4.2.0
python-decouple==3.8
Pillow>=10.4.0
reportlab>=4.2.2
psycopg2-binary>=2.9.9
python-barcode>=0.15.1
qrcode>=7.4.2
requests>=2.32.3
numpy>=2.1.0
